% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sinasc.R
\name{download.sinasc}
\alias{download.sinasc}
\title{Download SINASC}
\usage{
download.sinasc(
  inicio,
  fim = as.numeric(format(Sys.Date(), "\%Y")) - 1,
  UF,
  cod_mat = ""
)
}
\arguments{
\item{inicio}{The year that start the files extract.}

\item{fim}{The year that finish the files extract. By default the last year. Can't be the current year.}

\item{UF}{The state acronym.}

\item{cod_mat}{The birth establishment code.}
}
\value{
A data frame with the filtered raw SINASC data.
}
\description{
Download data about live birth from SINASC - DATASUS and tranform from .dbc file to data frame
}
\examples{
pe <- download.sinasc(2022,UF="PE")
sp <- download.sinasc(2020,2022,"SP")
}
